+++ types/CHANGELOG.md
@@ -21,9 +21,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## 1.0.0-rc.8 - 2023-XX-XX
 
+### Added
+
+- `{AliasId, NftId}::to_bech32`;
+- `Address::{is_valid_bech32, try_from_bech32_with_hrp}`;
+
 ### Changed
 
 - `OutputIdsResponse::items` from `Vec<String>` to `Vec<OutputId>`;
+- `Address::try_from_bech32` doesn't return the HRP anymore;
 
 ## 1.0.0-rc.7 - 2023-03-09