+++ wallet/CHANGELOG.md
@@ -21,6 +21,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## 1.0.0-rc.7 - 2023-XX-XX
 
+### Added
+
+- `AddressWrapper::try_from_bech32`;
+
 ### Changed
 
 - Resync outputs if a transaction got confirmed between syncing outputs and pending transactions to prevent not having unspent outputs afterwards;